+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Changelog
 
 ## [x]
+- Only show in certain places config
 - Delete function
 - ILIAS 5.4 support
 - Drop ILIAS 5.1 and ILIAS 5.2 support